\name{batch_config}
\alias{batch_config}
\title{Create a configuration file for batch downloading / clustering od sequences}
\description{
Creats a txt file in which download parametwers, like databases, marker name and clustering identity can be configured. After editing the configuration file and setting up the csv or table listing the taxonomic groups the primers should be optimised for, run the comand \code{\link{batch_download}}}
\usage{
batch_config(file)
}
\arguments{
  \item{file}{Give a file name for the config.txt}
}
\details{
}
\value{

}
\references{
}
\author{
}
\note{
}

\seealso{
\code{\link{batch_download}}}
\examples{
}
