\name{batch_download}
\alias{batch_download}
\title{
Batch download and process files
}
\description{
}
\usage{
batch_download(table, download = T, BOLD = T, GB = T, mito = T, clipping = 26)
}
\arguments{
  \item{table}{
}
  \item{download}{
}
  \item{BOLD}{
}
  \item{GB}{
}
  \item{mito}{
}
  \item{clipping}{
}
}
\details{
}
\value{}
\references{
}
\author{
}
\note{
}


\seealso{
}
\examples{
}
