% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differing_bases.R
\name{differing_bases}
\alias{differing_bases}
\title{Find positions where prey DNA differs from that of its predator, to aid in designing dietary primers. To be used after aligning sequences, as per https://www.youtube.com/watch?v=EcdTxGmcsEM}
\usage{
differing_bases(files = c("List of files"), predator_name,
  min_prey = 0.75, max_pred = 0.25, output = "plot")
}
\arguments{
\item{files}{a vector of input files}

\item{min_prey}{the minimum proportion of prey taxa which need to have a given base for it to be considered a potentially useful site for a primer}

\item{max_prey}{the maximum proportion of predator sequences which need to have a given base for it to be considered a potentially useful site for a primer}
}
\value{
Either a plot summarising the results, or a list of tibbles with useful values
}
\description{
Find positions where prey DNA differs from that of its predator, to aid in designing dietary primers. To be used after aligning sequences, as per https://www.youtube.com/watch?v=EcdTxGmcsEM
}
